--- src/NumSquareForm.tsx
@@ -1,31 +1,0 @@
-import React, { useState } from "react";
-import { Form } from "react-bootstrap";
-/*
-export function NumSquareForm(): JSX.Element {
-    return (
-        <div>
-            <Form.Group
-                controlId="numSquareForm"
-                style={{
-                    position: "relative",
-                    width: "130px",
-                    right: 0,
-                    bottom: -10
-                }}
-            >
-                <Form.Label>Number of Tanks:</Form.Label>
-                <Form.Control
-                    type="number"
-                    value={numSquares}
-                    onChange={(event: React.ChangeEvent<HTMLInputElement>) =>
-                        setNumSquares(parseInt(event.target.value))
-                    }
-                />
-            </Form.Group>
-        </div>
-    );
-}
-
-export default NumSquareForm;
-
-*/
