+++ src/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import ReactDOM from "react-dom";
 import Board from "./Board";
 import "./index.css";
@@ -40,3 +40,36 @@ observe((picPosition: [number, number]) => {
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
+
+/*
+import React, { useState } from "react";
+import { Form } from "react-bootstrap";
+
+export function NumSquareForm(): JSX.Element {
+    return (
+        <div>
+            <Form.Group
+                controlId="numSquareForm"
+                style={{
+                    position: "relative",
+                    width: "130px",
+                    right: 0,
+                    bottom: -10
+                }}
+            >
+                <Form.Label>Number of Tanks:</Form.Label>
+                <Form.Control
+                    type="number"
+                    value={numSquares}
+                    onChange={(event: React.ChangeEvent<HTMLInputElement>) =>
+                        setNumSquares(parseInt(event.target.value))
+                    }
+                />
+            </Form.Group>
+        </div>
+    );
+}
+
+export default NumSquareForm;
+
+*/
